api/requirements.txt
fastapi
uvicorn
pydantic
huggingface_hub
pyjwt
httpx
python-multipart